axios.get request set up currently for specific product_id endpoint
