+++ server/index.js
@@ -9,8 +9,8 @@ app.use(express.static(path.join(__dirname, '../client/public')));
 // app.use(express.urlencoded());
 
 app.get('/', (req, res) => {
-  res.send('Server running!')
-})
+  res.send('Server running!');
+});
 
 app.listen(port, () => {
   console.log(`server listening on port ${port}`);
